+++ packages/modules/cra-template-web-viewer/template/src/AuthorizationOptions.ts
@@ -1,0 +1,23 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import { WebAuthorizationOptions } from "@itwin/web-viewer-react";
+
+export const AuthorizationOptions: WebAuthorizationOptions = {
+  config: {
+    scope: process.env.IMJS_AUTH_CLIENT_SCOPES ?? "",
+    clientId: process.env.IMJS_AUTH_CLIENT_CLIENT_ID ?? "",
+    redirectUri: process.env.IMJS_AUTH_CLIENT_REDIRECT_URI ?? "",
+    postSignoutRedirectUri: process.env.IMJS_AUTH_CLIENT_LOGOUT_URI,
+    responseType: "code",
+    authority: process.env.IMJS_AUTH_AUTHORITY,
+  },
+};
+
+export default AuthorizationOptions;
