--- packages/modules/cra-template-web-viewer/template/src/Header.module.scss
@@ -1,16 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-.header {
-  background-color: #282c34;
-
-  .buttonContainer {
-    text-align: center;
-
-    button {
-      width: 100px;
-      margin: 10px !important;
-    }
-  }
-}
